requirements.txt
# requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
plotly==6.1.2
langchain-core==0.1.22
langchain-google-genai==0.0.7
google-generativeai==0.3.1
supabase==2.3.0
extra-streamlit-components>=0.1.60  # 用於處理認證狀態